backend/requirements.txt
Django>=4.2,<5
requests>=2.31
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
redis>=5.0
django-redis>=5.4
django-environ>=0.11
Pillow>=10.0
boto3>=1.33
django-storages[google]>=1.14
django-filter>=23.5
whitenoise>=6.6,<7
openai>=1.40
pgvector>=0.3
